Updates license header code templates
Updates the code templates with the license header.
The templates refer to the current year instead of a fixed year.
The templates refer to Alliander N.V. instead of Smart Society Services.

Reformatted the lines to match those shown at the Apache URL, and
removed the indentation before the URL in the template, as it is prone
to be removed upon formatting the code in the editor, which is
inconvenient.