Adds initial conventions for XML formatting
Describes some initial conventions agreed upon to be used for XML
formatting. Establishing these conventions is a work in progress and
this is a very limited starting point.

Removes screen shots for save actions. The configuration for these
actions is updated as described in the readme.

Updates indentation of nested bullets.